# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/enrollments": {
      "post": {
        "description": "RN34–RN38. O controle de vagas é seguro sob requisições simultâneas.",
        "operationId": "Enrollments_enroll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`. Perfil diferente de aluno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Curso inexistente ou em rascunho",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflito de estado — `ALREADY_ENROLLED`, `COURSE_FULL`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "ALREADY_ENROLLED": {
                    "summary": "ALREADY_ENROLLED",
                    "value": {
                      "statusCode": 409,
                      "code": "ALREADY_ENROLLED",
                      "message": "Você já está matriculado neste curso",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  },
                  "COURSE_FULL": {
                    "summary": "COURSE_FULL",
                    "value": {
                      "statusCode": 409,
                      "code": "COURSE_FULL",
                      "message": "Não há vagas disponíveis neste curso",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada — `COURSE_NOT_OPEN`. Curso arquivado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "COURSE_NOT_OPEN": {
                    "summary": "COURSE_NOT_OPEN",
                    "value": {
                      "statusCode": 422,
                      "code": "COURSE_NOT_OPEN",
                      "message": "Este curso não está aberto para matrículas",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aluno se matricula em um curso publicado com vagas",
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/api/enrollments/me": {
      "get": {
        "description": "Mais recentes primeiro. Inclui matrículas canceladas, a menos que filtre por status.",
        "operationId": "Enrollments_mine",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "CANCELLED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Matrículas do aluno logado (com progresso, notas e situação)",
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/api/enrollments/{id}": {
      "get": {
        "description": "Aluno dono, professor do curso ou admin.",
        "operationId": "Enrollments_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Matrícula inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe e boletim da matrícula",
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/api/enrollments/{id}/cancel": {
      "post": {
        "description": "RN39–RN40. Depois de cancelar, o aluno pode se matricular de novo.",
        "operationId": "Enrollments_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`. Matrícula de outro aluno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Matrícula inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada — `ENROLLMENT_NOT_ACTIVE`, `ENROLLMENT_HAS_GRADES`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "ENROLLMENT_NOT_ACTIVE": {
                    "summary": "ENROLLMENT_NOT_ACTIVE",
                    "value": {
                      "statusCode": 422,
                      "code": "ENROLLMENT_NOT_ACTIVE",
                      "message": "Esta matrícula já está cancelada",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  },
                  "ENROLLMENT_HAS_GRADES": {
                    "summary": "ENROLLMENT_HAS_GRADES",
                    "value": {
                      "statusCode": 422,
                      "code": "ENROLLMENT_HAS_GRADES",
                      "message": "Matrículas com notas lançadas não podem ser canceladas",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancela matrícula ativa sem notas",
        "tags": [
          "Enrollments"
        ]
      }
    },
...
    "/api/courses/{id}/enrollments": {
      "get": {
        "description": "Professor dono ou admin. Ativas primeiro, depois as canceladas.",
        "operationId": "Enrollments_byCourse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Curso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Turma do curso",
        "tags": [
          "Enrollments"
        ]
      }
    },
...
  },
  "info": {
    "title": "EduTrack API",
    "description": "API da plataforma EduTrack: cursos, aulas, matrículas, progresso, notas e certificados.\n\n### Autenticação\n1. `POST /api/auth/login` com `aluno@edutrack.dev` / `Senha@123` (ou outro usuário do seed).\n2. Copie o `accessToken` e clique em **Authorize**.\n\nRotas marcadas como **públicas** aceitam token opcional (com token, a resposta traz o contexto do usuário).\n\n### Erros\nSempre no formato `ErrorResponseDto`: `{ statusCode, code, message, details?, path, timestamp, requestId }`.\nFaça asserções pelo **`code`** (estável); `message` é texto para humanos.\n\n| Status | Significado |\n|---|---|\n| 400 | Dados inválidos (`details` lista os campos) |\n| 401 | Não autenticado / credenciais inválidas |\n| 403 | Autenticado, mas sem permissão |\n| 404 | Não existe ou não é visível para você |\n| 409 | Conflito de estado ou unicidade |\n| 422 | Regra de negócio violada |\n| 429 | Limite de tentativas de login |\n\nToda resposta traz o header `X-Request-Id`, que também aparece no log do servidor.\n\nAs regras citadas como RNxx estão em `docs/regras-de-negocio.md`.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "accessToken obtido no login"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 422
          },
          "code": {
            "description": "Código estável para uso por máquina (frontend e testes)",
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "Adicione ao menos uma aula antes de publicar",
            "description": "Texto para humanos; pode mudar"
          },
          "details": {
            "description": "Presente apenas em erros de validação (400)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/courses/8f0c.../publish"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Mesmo valor do header X-Request-Id"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "CreateEnrollmentDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "courseId"
        ]
      },
...
    }
  }
}